DataScience-2/ex00/requirements.txt
pandas
psycopg2
sqlalchemy
matplotlib
python-dotenv